# Fix cases which pending items flowing will not be accounted after an error