--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "caminho",
-  "version": "1.7.1",
+  "version": "1.7.2",
   "description": "Tool for creating efficient data pipelines in a JavaScript environment",
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm5/index.js",