--- a/src/operators/filter.ts
+++ b/src/operators/filter.ts
@@ -1,7 +1,7 @@
 import { filter as filterRxJs } from 'rxjs'
 import { type PendingDataControl } from '../utils/PendingDataControl'
 import { type Loggers, type ValueBag } from '../types'
-import { type OperatorApplier } from './helpers/operatorHelpers'
+import { type OperatorApplierWithRunId } from './helpers/operatorHelpers'
 
 export type FilterPredicate = (valueBag: ValueBag, index: number) => boolean
 
@@ -9,14 +9,14 @@ export function filter(
   predicate: FilterPredicate,
   loggers: Loggers,
   pendingDataControl?: PendingDataControl,
-): OperatorApplier {
-  function wrappedFilter(valueBag: ValueBag, index: number): boolean {
+): OperatorApplierWithRunId {
+  function wrappedFilter(valueBag: ValueBag, index: number, runId: string): boolean {
     const startedAt = new Date()
     loggers.onStepStarted([valueBag])
     try {
       const filterResult = predicate(valueBag, index)
       if (!filterResult) {
-        pendingDataControl?.decrement()
+        pendingDataControl?.decrement(runId)
       }
       loggers.onStepFinished([valueBag], startedAt)
       return filterResult
@@ -26,5 +26,7 @@ export function filter(
     }
   }
 
-  return filterRxJs(wrappedFilter)
+  return function filterOperatorWithRunId(runId: string) {
+    return filterRxJs((valueBag, index) => wrappedFilter(valueBag, index, runId))
+  }
 }