--- a/src/utils/PendingDataControl.ts
+++ b/src/utils/PendingDataControl.ts
@@ -1,17 +1,29 @@
 export class PendingDataControlInMemory implements PendingDataControl {
   public size = 0
+  private buckets = new Map()
 
-  increment(value = 1): void {
+  increment(bucketId: string, value = 1): void {
     this.size += value
+    const current = this.buckets.get(bucketId) ?? 0
+    this.buckets.set(bucketId, current + value)
   }
 
-  decrement(value = 1): void {
+  decrement(bucketId: string, value = 1): void {
     this.size -= value
+    const current = this.buckets.get(bucketId)
+    this.buckets.set(bucketId, current - value)
+  }
+
+  destroyBucket(bucketId: string) {
+    const inBucket = this.buckets.get(bucketId) ?? 0
+    this.size -= inBucket
+    this.buckets.delete(bucketId)
   }
 }
 
 export type PendingDataControl = {
   size: number
-  increment: (value?: number) => void
-  decrement: (value?: number) => void
+  increment: (bucketId: string, value?: number) => void
+  decrement: (bucketId: string, value?: number) => void
+  destroyBucket: (bucketId: string) => void
 }