--- a/src/utils/generateId.ts
+++ b/src/utils/generateId.ts
@@ -1,0 +1,3 @@
+export function generateId() {
+  return `${Date.now().toString(36)}-${Math.random().toString(36).substr(2, 9)}`
+}